Tweak Doxygen config to reduce the number of undocumented warnings
